requirements.txt
streamlit
pandas
openpyxl
python-dotenv
jinja2
email-validator
tqdm
sqlite-utils 